Fix low priority bugs: help text, terminal size guard, compiler warnings
- Fix help text "TAB/1-6" to "TAB/1-7" for 7 tabs (main.c)
- Add minimum terminal size check (20x8) at startup and on resize (ui.c)
- Fix mc_error_t signed comparison with explicit cast (error.c)
- Remove unused entry_count/news_count params from draw_status (ui.c)

+++ backend/src/error.c
@@ -17,7 +17,7 @@ static const char *error_strings[] = {
 
 const char *mc_error_str(mc_error_t err)
 {
-    if (err < 0 || err > MC_ERR_NOT_FOUND)
+    if ((int)err < 0 || err > MC_ERR_NOT_FOUND)
         return "Unknown error";
     return error_strings[err];
 }
